Pop up werkt

+++ index.php
@@ -51,8 +51,7 @@
                                 <td><?php echo $appointment['date'] ?></td>
                                 <td class="actions">
                                     <a href="update.php?id=<?= $appointment['Id'] ?>">Bewerken</a>
-                                    <a href="delete.php?id=<?= $appointment['Id'] ?>" >Verwijderen</a>
-                                </td>
+                                    <a href="delete.php?id=<?= $appointment['Id'] ?>" onclick="return confirm('Weet je zeker dat je dit record wilt verwijderen?')">Verwijderen</a>
                             </tr>
                         <?php endforeach; ?>
                     </tbody>
